component colors worked on
