+++ adov_flutter/lib/screens/home_screen/conference_stats.dart
@@ -19,10 +19,7 @@ class ConferenceStats extends StatelessWidget {
       child: Container(
         width: 400,
         margin: EdgeInsets.symmetric(horizontal: 20),
-        decoration: BoxDecoration(
-          color: Colors.grey,
-          borderRadius: BorderRadius.circular(15),
-        ),
+        decoration: ComponentBoxStyle.create(radius: 15),
         child: Column(
             mainAxisAlignment: MainAxisAlignment.spaceEvenly,
             crossAxisAlignment: CrossAxisAlignment.start,
@@ -41,10 +38,7 @@ class ConferenceStats extends StatelessWidget {
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
                   Container(
-                    decoration: BoxDecoration(
-                      color: Color.fromRGBO(128, 128, 128, 1),
-                      borderRadius: BorderRadius.circular(7.5),
-                    ),
+                    decoration: ComponentBoxStyle.create(radius: 7.5, color: InnerComponentColor),
                     height: 125,
                     width: 160,
                     child: Column(
@@ -65,10 +59,7 @@ class ConferenceStats extends StatelessWidget {
                   ),
                   Container(
                     margin: EdgeInsets.only(bottom: 20),
-                    decoration: BoxDecoration(
-                      color: Color.fromRGBO(128, 128, 128, 1),
-                      borderRadius: BorderRadius.circular(7.5),
-                    ),
+                    decoration: ComponentBoxStyle.create(radius: 7.5, color: InnerComponentColor),
                     height: 125,
                     width: 160,
                     child: Column(
