+++ adov_flutter/lib/style.dart
@@ -1,3 +1,5 @@
+import 'dart:ffi';
+
 import 'package:flutter/material.dart';
 
 const LargeTextSize = 22.0;
@@ -14,8 +16,60 @@ const Color TextColorFaint = Color.fromRGBO(125, 125, 125, 1.0);
 
 const Color AccentColor = Color(0xFF29335C);
 const Color ComponentColor = Color(0xFFB9BFDF);
+const Color InnerComponentColor = Color(0xFF818BC5);
 
 const DefaultPaddingHorizontal = 12.0;
+
+class ComponentBoxStyle {
+  static create({double radius: 0, Color color: ComponentColor}) {
+    return BoxDecoration(
+      color: color,
+      borderRadius: BorderRadius.circular(radius),
+    );
+  }
+}
+
+class CustomInputDecorator extends InputDecoration {
+  static InputDecoration decorator() {
+    return InputDecoration(
+      focusedErrorBorder: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(5.0),
+        borderSide: BorderSide(
+          color: Colors.red,
+          width: 1.5,
+        ),
+      ),
+      labelText: "Talk Date",
+      fillColor: Colors.white,
+      focusedBorder: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(5.0),
+        borderSide: BorderSide(
+          color: AccentColor,
+          width: 1.5,
+        ),
+      ),
+      enabledBorder: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(5.0),
+        borderSide: BorderSide(
+          color: Colors.grey,
+          width: 1.5,
+        ),
+      ),
+      errorBorder: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(5.0),
+        borderSide: BorderSide(
+          color: Colors.red,
+          width: 1.5,
+        ),
+      ),
+      errorStyle: TextStyle(
+        height: 0,
+        color: Colors.red
+      ),
+
+    );
+  }
+}
 
 const CardsTitleStyle = TextStyle(
   fontFamily: FontNameDefault,
